clippy fix and WDT

+++ src/drivers/axp.rs
@@ -1,8 +1,7 @@
 // src/drivers/axp.rs
 use anyhow::Result;
-use axp20x::{Axpxx, Power, PowerState};
+use axp20x::Axpxx;
 use crate::manager::I2cManager;
-use esp_idf_hal::delay::FreeRtos;
 
 pub struct PowerManager;
 
@@ -14,7 +13,7 @@ impl PowerManager {
     pub fn get_battery_percentage(&mut self, i2c: &mut I2cManager) -> Result<u8> {
         let mut axp = Axpxx::new(i2c);
         let percentage = axp.get_battery_percentage().map_err(|e| anyhow::anyhow!("{:?}", e))?;
-        Ok(percentage as u8)
+        Ok(percentage)
     }
 
     pub fn read_voltage(&mut self, i2c: &mut I2cManager) -> Result<u16> {
